Rework loggout

+++ client/src/layouts/layout.js
@@ -19,12 +19,22 @@ export default {
                 const graphqlQuery = {
                     query: `mutation{logout}`,
                 };
-                const loggoutRequest = await apiService.post(`${location.origin}/graphql`, graphqlQuery);
-                const loggoutData = loggoutRequest.data.data.logout;
-                if (loggoutData) {
-                    this.connected = false;
-                    localStorage.clear();
-                    this.$router.push('/connexion')
+                // const loggoutRequest = await apiService.post(`${location.origin}/graphql`, graphqlQuery);
+                const loggoutRequest = await apiService.post(`http://localhost:3000/graphql`, graphqlQuery);
+                const loggoutErrors  = loggoutRequest.data.errors;
+
+                if (loggoutErrors) {
+                    this.flashMessage.error({
+                        title: loggoutErrors[0].message,
+                        time: 8000,
+                    })
+                } else {
+                    const loggoutData = loggoutRequest.data.data.logout;
+                    if (loggoutData) {
+                        this.connected = false;
+                        localStorage.clear();
+                        this.$router.push('/connexion')
+                    }
                 }
             } catch (error) {
                 this.connected = false;
